Refactored UserController
